+++ plugins/11_TheWarWithin/localization/enUS.lua
@@ -610,3 +610,9 @@ L['discarded_can_suffix'] = 'cans kicked'
 L['options_icons_scrap_rewards'] = 'S.C.R.A.P. Heaps'
 L['options_icons_scrap_rewards_desc'] = 'Display rewards for {object:S.C.R.A.P. Heaps}.'
 L['scrap_heap_suffix'] = 'S.C.R.A.P. jobs completed'
+
+L['options_icons_youre_my_friend_now'] = '{achievement:41708}'
+L['options_icons_youre_my_friend_now_desc'] = 'Display {npc:239248} locations for {achievement:41708}.'
+L['rat_in_building_rail'] = 'Inside the building on the rail.'
+L['rat_2nd_floor_trash_can'] = 'Inside the building on the second floor on the trash can.'
+