+++ plugins/11_TheWarWithin/localization/frFR.lua
@@ -611,3 +611,8 @@ L['discarded_can_suffix'] = 'canettes frappées'
 L['options_icons_scrap_rewards'] = 'Tas de D.E.B.R.I.S.'
 L['options_icons_scrap_rewards_desc'] = 'Afficher les récompenses liées aux {object:Tas de D.E.B.R.I.S.}.'
 L['scrap_heap_suffix'] = 'Missions de D.E.B.R.I.S. terminées'
+
+L['options_icons_youre_my_friend_now'] = '{achievement:41708}'
+L['options_icons_youre_my_friend_now_desc'] = nil
+L['rat_in_building_rail'] = nil
+L['rat_2nd_floor_trash_can'] = nil